all done in my mind

+++ src/mouse.c
@@ -395,7 +395,7 @@ bool smoothlyMoveMouse(MMPoint endPoint,double speed)
 
 		/* Make sure we are in the screen boundaries!
 		 * (Strange things will happen if we are not.) */
-		if (pos.x >= screenSize.width || pos.y >= screenSize.height) {
+		if (pos.x >= (int32_t)screenSize.width || pos.y >= (int32_t)screenSize.height) {
 			return false;
 		}
